Moved all dockering to docker_helper
Also, fixed a small bug where the database container would
not be stopped if a test unsuccessfully started and returned.
